refactor: Move Python Binding Test Setup Back To Bindings Crate
