+++ rust/python-bindings/benches/performance.rs
@@ -8,6 +8,8 @@
     reason = "criterion_group generates an undocumented entrypoint, and benchmarks fail fast with unwrap during setup"
 )]
 
+use std::sync::OnceLock;
+
 use avdschema::Load as _;
 use avdschema::Store;
 use criterion::Criterion;
@@ -19,6 +21,27 @@ use python_bindings::bindings;
 use test_schema_store::get_store_gz_path;
 
 const TEST_DATA: &str = r#"{"fabric_name":"foo","type":"l3ls-evpn"}"#;
+
+static INIT_PY: OnceLock<()> = OnceLock::new();
+static INIT_STORE: OnceLock<()> = OnceLock::new();
+
+fn setup_python_with_store() {
+    INIT_PY.get_or_init(|| {
+        pyo3::append_to_inittab!(bindings);
+        pyo3::Python::initialize();
+    });
+    INIT_STORE.get_or_init(|| {
+        pyo3::Python::attach(|py| {
+            let module = py.import("_bindings").unwrap();
+            let kwargs = PyDict::new(py);
+            let file = py.detach(get_store_gz_path);
+            kwargs.set_item("file", file).unwrap();
+            module
+                .call_method("init_store_from_file", (), Some(&kwargs))
+                .unwrap();
+        });
+    });
+}
 
 fn benchmark_load_and_resolve_store(criterion: &mut Criterion) {
     let schema_file = get_store_gz_path();
@@ -34,7 +57,7 @@ fn benchmark_load_and_resolve_store(criterion: &mut Criterion) {
 }
 
 fn benchmark_get_validated_data(criterion: &mut Criterion) {
-    test_schema_store::setup_python_with_store(|| pyo3::append_to_inittab!(bindings));
+    setup_python_with_store();
     criterion.bench_function("get_validated_data", |bencher| {
         pyo3::Python::attach(|py| {
             let module = py.import("_bindings").unwrap();
