+++ rust/python-bindings/src/tests/mod.rs
@@ -2,19 +2,39 @@
 // Use of this source code is governed by the Apache License 2.0
 // that can be found in the LICENSE file.
 
+use std::sync::OnceLock;
+
 use pyo3::types::PyAnyMethods as _;
+use pyo3::types::PyDict;
 
 use crate::bindings;
 
 mod passwords;
 mod validation;
 
+static INIT_PY: OnceLock<()> = OnceLock::new();
+static INIT_STORE: OnceLock<()> = OnceLock::new();
+
 fn setup_python() {
-    test_schema_store::setup_python(|| pyo3::append_to_inittab!(bindings));
+    INIT_PY.get_or_init(|| {
+        pyo3::append_to_inittab!(bindings);
+        pyo3::Python::initialize();
+    });
 }
 
 fn setup() {
-    test_schema_store::setup_python_with_store(|| pyo3::append_to_inittab!(bindings));
+    setup_python();
+    INIT_STORE.get_or_init(|| {
+        pyo3::Python::attach(|py| {
+            let module = py.import("_bindings").unwrap();
+            let kwargs = PyDict::new(py);
+            let file = py.detach(test_schema_store::get_store_gz_path);
+            kwargs.set_item("file", file).unwrap();
+            module
+                .call_method("init_store_from_file", (), Some(&kwargs))
+                .unwrap();
+        });
+    });
 }
 
 fn get_path_and_message_from_py_violation(
