+++ rust/python-bindings/src/tests/passwords/test_simple_7.rs
@@ -4,7 +4,7 @@
 
 use pyo3::types::PyAnyMethods as _;
 
-use super::super::setup_python;
+use crate::tests::setup_python;
 
 #[test]
 fn simple_7_encrypt_py_ok() {
