+++ rust/test_schema_store/Cargo.toml
@@ -5,7 +5,6 @@ edition = "2024"
 [workspace]
 
 [dependencies]
-pyo3 = { version = "0.29.0", features = ["abi3-py310"] }
 reqwest = { version = "0.12.24", default-features = false, features = [
     "blocking",
     "default-tls",
